Add Script on TableInfo and add a getScript public method on Script manager.
Minor fix the timestamp fields in the scripts.

+++ src/javatests/com/google/cloud/bigquery/dwhassessment/extractiontool/faketd/teradata_tables.sql
@@ -17,7 +17,7 @@ CREATE SCHEMA DBC;
 CREATE TABLE DBC.TablesV (
   DatabaseName VARCHAR(128),
   TableName VARCHAR(128),
-  AccessCount INTEGER,
+  AccessCount BIGINT,
   LastAccessTimeStamp TIMESTAMP(0),
   LastAlterTimeStamp TIMESTAMP(0),
   TableKind CHAR(1),
